Improve components hierarchy
